HHH-18271 Avoid query validations of cached queries by doing validation eagerly. Cache allowed result types per query interpretation
