+++ hibernate-core/src/main/java/org/hibernate/query/spi/SimpleHqlInterpretationImpl.java
@@ -6,8 +6,12 @@
  */
 package org.hibernate.query.spi;
 
+import java.util.concurrent.ConcurrentHashMap;
+
 import org.hibernate.query.sqm.internal.DomainParameterXref;
+import org.hibernate.query.sqm.internal.SqmUtil;
 import org.hibernate.query.sqm.tree.SqmStatement;
+import org.hibernate.query.sqm.tree.select.SqmSelectStatement;
 
 /**
  * @author Steve Ebersole
@@ -16,6 +20,7 @@ public class SimpleHqlInterpretationImpl<R> implements HqlInterpretation<R> {
 	private final SqmStatement<R> sqmStatement;
 	private final ParameterMetadataImplementor parameterMetadata;
 	private final DomainParameterXref domainParameterXref;
+	private final ConcurrentHashMap<Class<?>, Object> allowedReturnTypes;
 
 	public SimpleHqlInterpretationImpl(
 			SqmStatement<R> sqmStatement,
@@ -24,6 +29,7 @@ public class SimpleHqlInterpretationImpl<R> implements HqlInterpretation<R> {
 		this.sqmStatement = sqmStatement;
 		this.parameterMetadata = parameterMetadata;
 		this.domainParameterXref = domainParameterXref;
+		this.allowedReturnTypes = new ConcurrentHashMap<>();
 	}
 
 	@Override
@@ -39,5 +45,16 @@ public class SimpleHqlInterpretationImpl<R> implements HqlInterpretation<R> {
 	@Override
 	public DomainParameterXref getDomainParameterXref() {
 		return domainParameterXref.copy();
+	}
+
+	@Override
+	public void validateResultType(Class<?> resultType) {
+		assert sqmStatement instanceof SqmSelectStatement<?>;
+		if ( resultType != null && !SqmUtil.isResultTypeAlwaysAllowed( resultType ) ) {
+			if ( !allowedReturnTypes.containsKey( resultType ) ) {
+				SqmUtil.checkQueryReturnType( ( (SqmSelectStatement<R>) sqmStatement ).getQueryPart(), resultType );
+				allowedReturnTypes.put( resultType, Boolean.TRUE );
+			}
+		}
 	}
 }
